requirements.txt
django==3.0.5
django-bootstrap-form==3.4
djangorestframework==3.11.0
markdown==3.2.1
django-filter==2.2.0
django-widget-tweaks==1.4.8
drf-yasg==1.17.1